Fix self-review nits
