+++ beacon-chain/state/state-native/beacon_state_mainnet.go
@@ -10,6 +10,7 @@ import (
 	customtypes "github.com/prysmaticlabs/prysm/v3/beacon-chain/state/state-native/custom-types"
 	nativetypes "github.com/prysmaticlabs/prysm/v3/beacon-chain/state/state-native/types"
 	"github.com/prysmaticlabs/prysm/v3/beacon-chain/state/stateutil"
+	"github.com/prysmaticlabs/prysm/v3/consensus-types/interfaces"
 	eth2types "github.com/prysmaticlabs/prysm/v3/consensus-types/primitives"
 	ethpb "github.com/prysmaticlabs/prysm/v3/proto/prysm/v1alpha1"
 )
@@ -44,7 +45,7 @@ type BeaconState struct {
 	inactivityScores             []uint64                         `ssz-gen:"true" ssz-max:"1099511627776"`
 	currentSyncCommittee         *ethpb.SyncCommittee             `ssz-gen:"true"`
 	nextSyncCommittee            *ethpb.SyncCommittee             `ssz-gen:"true"`
-	latestExecutionPayloadHeader nativetypes.ExecutionPayloadHeader `ssz-gen:"true"`
+	latestExecutionPayloadHeader interfaces.ExecutionPayloadHeader `ssz-gen:"true"`
 
 	lock                  sync.RWMutex
 	dirtyFields           map[nativetypes.FieldIndex]bool
